+++ src/main/java/com/smu/tkk/repository/TradeChatRoomRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface TradeChatRoomRepository extends JpaRepository<TradeChatRoom,Integer> {
+public interface TradeChatRoomRepository extends JpaRepository<TradeChatRoom,Long> {
     // 내가 속한 채팅방 목록
     Page<TradeChatRoom> findByMemberId(Long memberId, Pageable pageable);
