+++ src/main/java/com/smu/tkk/repository/TradePostRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.data.jpa.repository.Modifying;
 import org.springframework.data.jpa.repository.Query;
 
-public interface TradePostRepository extends JpaRepository<TradePost,Integer> {
+public interface TradePostRepository extends JpaRepository<TradePost,Long> {
     //거래 상태 변경
     @Modifying
     @Query(value = "update TradePost t set t.status=:status where t.id=:postId")
